Two explicit doors: Add a folder above the GitHub picker
The one-box "paste a URL or a path" input asked users to think in
Flow's taxonomy instead of their own nouns. Replaced with two doors:
"Add a folder" (path in → if it contains GitHub repos, one dialog asks
the base branch for each and adds them as normal GitHub-connected
repos with the folder as work surface; docs remainder subordinate)
sitting directly above the GitHub repos picker (the existing checklist,
extracted to components/RepoPicker.tsx and now also on the home hero).
URL paste survives only as a small "or paste a repo URL" expander in
the GitHub section. AddSource.tsx deleted; shared plumbing in
components/sources/kit.tsx. Prod mode hides the folder door.

Verified: tsc/eslint/build clean, old combined input absent from all
compiled chunks, and both pages screenshot-reviewed on an isolated
instance (duplicate section heading and redundant hero copy caught
and fixed in review).
