+++ dashboard/src/app/page.tsx
@@ -52,12 +52,6 @@ interface AuditRow {
   created_at?: number;
 }
 
-interface GhRepo {
-  full_name: string;
-  url: string;
-  default_branch: string;
-}
-
 type HomeState = "loading" | "engine-down" | "no-brain" | "empty" | "building" | "alive";
 
 // ─── Humanized translations ────────────────────────────────────────────────────
@@ -249,213 +243,6 @@ function SourceCard({ icon, title, description, locked, lockReason, children, on
           {children}
         </div>
       )}
-    </div>
-  );
-}
-
-// ─── Repo Picker (shared between State 1 GitHubCard and Home GitHub management) ──
-
-interface RepoPickerPanelProps {
-  onConnected: () => void;
-  onClose?: () => void;
-}
-
-function RepoPickerPanel({ onConnected, onClose }: RepoPickerPanelProps) {
-  const [repos, setRepos] = useState<GhRepo[]>([]);
-  const [repoSrc, setRepoSrc] = useState<"gh_cli" | "pat" | "none">("none");
-  const [pat, setPat] = useState("");
-  const [selected, setSelected] = useState<Set<string>>(new Set());
-  const [branches, setBranches] = useState<Record<string, string>>({});
-  const [loading, setLoading] = useState(true);
-  const [saving, setSaving] = useState(false);
-  const [msg, setMsg] = useState("");
-
-  useEffect(() => {
-    setLoading(true);
-    fetch("/api/github/repos")
-      .then((r) => r.json())
-      .then((d: { source: "gh_cli" | "pat" | "none"; repos: GhRepo[] }) => {
-        setRepoSrc(d.source);
-        setRepos(d.repos ?? []);
-      })
-      .catch(() => {})
-      .finally(() => setLoading(false));
-  }, []);
-
-  async function handleSavePat(e: FormEvent) {
-    e.preventDefault();
-    setSaving(true);
-    try {
-      await fetch("/api/github/repos", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ action: "save_pat", pat }),
-      });
-      setPat("");
-      setLoading(true);
-      const r = await fetch("/api/github/repos");
-      const d = await r.json() as { source: "gh_cli" | "pat" | "none"; repos: GhRepo[] };
-      setRepoSrc(d.source);
-      setRepos(d.repos ?? []);
-    } finally {
-      setSaving(false);
-      setLoading(false);
-    }
-  }
-
-  async function handleAdd() {
-    if (selected.size === 0) return;
-    setSaving(true);
-    setMsg("");
-    const toAdd = repos
-      .filter((r) => selected.has(r.full_name))
-      .map((r) => ({
-        ...r,
-        branch: branches[r.full_name]?.trim() || r.default_branch,
-      }));
-    try {
-      const res = await fetch("/api/github/repos", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ action: "add_repos", repos: toAdd }),
-      });
-      const data = await res.json() as { ok: boolean; results?: Array<{ ok: boolean }> };
-      if (data.ok) {
-        const count = (data.results ?? []).filter((r) => r.ok).length;
-        setMsg(`${count} repo${count !== 1 ? "s" : ""} connected — indexing will start soon.`);
-        setSelected(new Set());
-        onConnected();
-      }
-    } catch {
-      setMsg("Something went wrong.");
-    } finally {
-      setSaving(false);
-    }
-  }
-
-  if (loading) {
-    return <p className="text-text-muted text-[13px]">Loading your repositories…</p>;
-  }
-
-  return (
-    <div className="flex flex-col gap-3">
-      {repoSrc === "none" ? (
-        <form onSubmit={handleSavePat} className="flex flex-col gap-3">
-          <p className="text-[13px] text-text-muted">
-            Paste a GitHub fine-grained PAT (repo read access) to list your repos.
-          </p>
-          <input
-            type="password"
-            value={pat}
-            onChange={(e) => setPat(e.target.value)}
-            placeholder="github_pat_…"
-            className="w-full rounded-md border border-line bg-cream px-4 py-2.5 text-[14px] text-text placeholder:text-text-muted/60 outline-none focus:border-ink/20"
-          />
-          <Button type="submit" disabled={!pat.trim() || saving} arrow>
-            {saving ? "Connecting…" : "Connect GitHub"}
-          </Button>
-        </form>
-      ) : (
-        <div className="flex flex-col gap-3">
-          <p className="text-[12px] text-text-muted">
-            {repos.length} repos available via {repoSrc === "gh_cli" ? "gh CLI" : "GitHub PAT"}.
-            Select to connect.
-          </p>
-          <div className="max-h-48 overflow-y-auto rounded-lg border border-line">
-            {repos.slice(0, 60).map((repo) => (
-              <label
-                key={repo.full_name}
-                className="flex items-center gap-3 px-3 py-2.5 cursor-pointer hover:bg-sand border-b border-line last:border-b-0 transition-colors"
-              >
-                <input
-                  type="checkbox"
-                  checked={selected.has(repo.full_name)}
-                  onChange={() => setSelected((prev) => {
-                    const next = new Set(prev);
-                    if (next.has(repo.full_name)) next.delete(repo.full_name);
-                    else next.add(repo.full_name);
-                    return next;
-                  })}
-                  className="w-3.5 h-3.5"
-                  style={{ accentColor: "var(--accent)" }}
-                />
-                <span
-                  style={{ fontFamily: "var(--font-mono)" }}
-                  className="text-[12px] text-text"
-                >
-                  {repo.full_name}
-                </span>
-                {selected.has(repo.full_name) && (
-                  <input
-                    type="text"
-                    value={branches[repo.full_name] ?? repo.default_branch}
-                    onChange={(e) =>
-                      setBranches((prev) => ({ ...prev, [repo.full_name]: e.target.value }))
-                    }
-                    onClick={(e) => e.preventDefault()}
-                    spellCheck={false}
-                    title="Branch to watch"
-                    style={{ fontFamily: "var(--font-mono)" }}
-                    className="ml-auto w-32 rounded border border-line bg-cream px-2 py-0.5 text-[11px] text-text outline-none focus:border-ink/20"
-                  />
-                )}
-              </label>
-            ))}
-          </div>
-          {selected.size > 0 && (
-            <Button onClick={handleAdd} disabled={saving} arrow>
-              {saving ? "Connecting…" : `Connect ${selected.size} repo${selected.size !== 1 ? "s" : ""}`}
-            </Button>
-          )}
-        </div>
-      )}
-      {msg && <p className="text-[12px] text-text-muted">{msg}</p>}
-      {onClose && (
-        <button
-          onClick={onClose}
-          className="self-start text-[11px] text-text-muted hover:text-ink transition mt-1"
-          style={{ fontFamily: "var(--font-mono)", textTransform: "uppercase", letterSpacing: "0.08em" }}
-        >
-          Cancel
-        </button>
-      )}
-    </div>
-  );
-}
-
-// ─── GitHub Card (State 1 — no repos yet) ────────────────────────────────────
-
-function GitHubCard({ onConnected }: { onConnected: () => void }) {
-  const [open, setOpen] = useState(false);
-
-  if (!open) {
-    return (
-      <SourceCard
-        icon={<BrandIcon name="github" size={20} />}
-        title="GitHub"
-        description="Pick repositories to understand."
-        onClick={() => setOpen(true)}
-      />
-    );
-  }
-
-  return (
-    <div className="rounded-xl border border-ink/20 bg-paper p-5 flex flex-col gap-4">
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-2">
-          <BrandIcon name="github" size={18} className="text-ink" />
-          <span style={{ fontFamily: "var(--font-display)" }} className="text-ink text-[15px] font-medium">
-            GitHub
-          </span>
-        </div>
-        <button
-          onClick={() => setOpen(false)}
-          className="text-text-muted hover:text-ink text-sm transition"
-        >
-          ✕
-        </button>
-      </div>
-      <RepoPickerPanel onConnected={onConnected} onClose={() => setOpen(false)} />
     </div>
   );
 }
@@ -1168,8 +955,6 @@ export default function HomePage() {
             <Kicker>Other ways to connect</Kicker>
           </div>
           <div className="grid gap-4 mt-3">
-            <GitHubCard onConnected={() => { loadAll(); setState("building"); }} />
-
             <ApiKeyCard
               icon={<BrandIcon name="linear" size={20} />}
               title="Linear"
